refactor: MovieController의 모든 메서드가 ResponseEntity<CustomResponse>를 반환하도록 변경

+++ src/main/java/hello/movie/dto/MovieDto.java
@@ -12,7 +12,7 @@ public class MovieDto {
     private String title;
     private String overview;
     private LocalDate releaseDate;
-    private float rating;
+    private Double rating;
     private String posterPath;
     private String trailerPath;
     private List<MovieGenreDto> genres;
